% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allwise_p_chisq.R
\name{allwise_p_chisq}
\alias{allwise_p_chisq}
\title{Calculate p-value for C-chisq}
\usage{
allwise_p_chisq(input, num_permute = 10000, na.rm = F)
}
\arguments{
\item{input}{numeric array}

\item{num_permute}{numeric}

\item{na.rm}{boolean}
}
\value{
length-one numeric
}
\description{
\code{allwise_p_chisq} returns the p-value for 'all lineages'
contrast using chi-square. Setting na.rm = T assumes that missing values
are true negatives.
}
\examples{
array1 <- array (0,c(5000,20))
array1[cbind(1:20,1:20)] <- 1
array1[1:5,1:3] <- 1
array1[6:10,4:6] <- 1
allwise_p_chisq (array1)

}
