% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_pa.R
\name{likelihood_pa}
\alias{likelihood_pa}
\title{Calculate log-likelihood scores for estimate_pa.R}
\usage{
likelihood_pa(test_prop_adaptive, input, na.rm = F)
}
\arguments{
\item{test_prop_adaptive}{numeric}

\item{input}{numeric vector}
}
\value{
length-one numeric
}
\description{
\code{likelihood_pa} Calculates the log-likelihood scores
for use in estimate_pa.R. Setting na.rm = T assumes that missing values
are true negatives.
}
\examples{
array1 <- array (0,c(5000,20))
array1[cbind(1:20,1:20)] <- 1
array1[1:5,1:3] <- 1
array1[6:10,4:6] <- 1
test_proportion <- 0.2
likelihood_pa (test_proportion, array1)

}
