% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_c_chisq.R
\name{pairwise_c_chisq}
\alias{pairwise_c_chisq}
\title{Calculate C-chisq}
\usage{
pairwise_c_chisq(input, num_permute = 10000, na.rm = F)
}
\arguments{
\item{input}{numeric array}

\item{num_permute}{numeric}

\item{na.rm}{boolean}
}
\value{
length-one numeric
}
\description{
\code{pairwise_c_chisq} returns the pairwise C-score using
the chi-square approach. Setting na.rm = T assumes that missing values
are true negatives.
}
\examples{
array1 <- array (0,c(5000,20))
array1[cbind(1:20,1:20)] <- 1
array1[1:5,1:3] <- 1
array1[6:10,4:6] <- 1
pairwise_c_chisq (array1)

}
