% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_c_hyper.R
\name{pairwise_c_hyper}
\alias{pairwise_c_hyper}
\title{Calculate C-hyper}
\usage{
pairwise_c_hyper(input, na.rm = F)
}
\arguments{
\item{input}{numeric array}

\item{na.rm}{boolean}
}
\value{
length-one numeric
}
\description{
\code{pairwise_c_hyper} returns the pairwise C-score using
the hypergeometric approach. Setting na.rm = T assumes that missing values
are true negatives.
}
\examples{
array1 <- array (0,c(5000,20))
array1[cbind(1:20,1:20)] <- 1
array1[1:5,1:3] <- 1
array1[6:10,4:6] <- 1
pairwise_c_hyper (array1)

}
