% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_p_chisq.R
\name{single_p_chisq}
\alias{single_p_chisq}
\title{Calculate p-value for C-chisq}
\usage{
single_p_chisq(input1, input2, num_permute = 10000, na.rm = F)
}
\arguments{
\item{input1}{numeric vector}

\item{input2}{numeric vector}

\item{num_permute}{numeric}

\item{na.rm}{boolean}
}
\value{
length-one numeric
}
\description{
\code{single_p_chisq} returns the p-value for a single pair
of lineages, contrasted using chi-square. Setting na.rm = T assumes that 
missing values are true negatives.
}
\examples{
vector1 <- rep (0,1000)
vector2 <- rep (0,1000)
vector1[c(1:5,15:20,24,31:35,47:50,75)] <- 1
vector2[c(6:9,12,19:23,28:31,78:80)] <- 1
single_p_chisq (vector1, vector2)

}
